fix: allow empty values for env vars by always inserting value (#80)
This commit allows empty values for env vars in order to make the behavior consistent with the normal command execution.

Suppose we set an env var called FOO with `export FOO=` (note this is an empty), and we have a simple script that prints this env var:

```ts
console.log(`Foo: ${Deno.env.get("FOO")}`);
```

Running this script via `deno run` gets us `Foo: `.
While via `deno task` (with `"foo": "deno run --allow-env foo.ts"` in deno.json) it gets us `Foo: undefined`.
This might be a significant difference if it matters whether the env var is set or not. So this PR fixes this behavior by always setting given env vars even if they are empty.